peg movements work
The pegs now move in sync with the algorithm.

This is a very rough alpha version of what will happen with a lot of
kruft. However, it provides the proof-of-concept for the idea. Now that
the pegs move, we can move the discs and fully emulate how the algorithm
works  directly (instead of hypothetically).

Pretty damn cool!